refactor customDashboard
