+++ components/organisms/fullCalendar/__tests__/stories.js
@@ -27,6 +27,10 @@ import { OffCanvas, Button, SelectField, FormLabel, FormField, CheckboxField } f
 
 import { FormForm } from 'index'
 
+import { CustomCalendarStyles } from './customCalendarStyles'
+
+export const main = () => <BaseComponent />
+
 export default {
   title: 'Organisms/Full Calendar',
   components: Calendar,
@@ -72,52 +76,10 @@ export const BaseComponent = (props = {}) => {
   return <Calendar ref={calendarRef} {...defaultCalendarProps} />
 }
 
-export const main = () => <BaseComponent />
-
 export const events = () => <BaseComponent events={Events} />
 
-export const calendarStyles = () => {
-  const [showOffCanvas, setShowOffCanvas] = useState(false)
-  const { errors, handleSubmit, register } = useForm()
-  const [calendarStyles, setCalendarStyles] = useState({})
-
-  const onSubmit = data => {
-    console.log('data submitted ', data)
-    const newCalendarStyles = {
-      height: Number(data.height)
-    }
-    setCalendarStyles(newCalendarStyles)
-    setShowOffCanvas(false)
-  }
-
-  const CalendarStyleForm = () => {
-    return (
-      <FormForm handleSubmit={handleSubmit(onSubmit)}>
-        <FormLabel label='Title'>
-          <FormField
-            type='number'
-            name='height'
-            errors={errors}
-            placeholder='Introduce the calendar height'
-            register={register}
-          />
-        </FormLabel>
-        <Button content='Submit' type='submit' />
-      </FormForm>
-    )
-  }
-
-  return (
-    <>
-      <button onClick={() => setShowOffCanvas(true)}>Open canvas</button>
-      {showOffCanvas && (
-        <OffCanvas headerText='Add custom event' show={showOffCanvas}>
-          {CalendarStyleForm()}
-        </OffCanvas>
-      )}
-      <BaseComponent {...calendarStyles} />
-    </>
-  )
+export const CustomCalendar = () => {
+  return <CustomCalendarStyles />
 }
 export const eventsWithEventAdditionOnClick = () => {
   const [events, setEvents] = useState(Events || [])
